Create adaptation of favorite page
